接口版本更新-->telegram-bot-api-v5

+++ config.go
@@ -2,7 +2,7 @@ package main
 
 type Config struct {
 	Token       string `json:"token" mapstructure:"token" yaml:"token"`
-	SuperUserId int    `json:"super_user_id" mapstructure:"super_user_id" yaml:"super_user_id"`
+	SuperUserId int64  `json:"super_user_id" mapstructure:"super_user_id" yaml:"super_user_id"`
 	Debug       bool   `json:"debug" mapstructure:"debug" yaml:"debug"`
 	ChanSize    int    `json:"chan_size" mapstructure:"chan_size" yaml:"chan_size"`
 }
